% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_roc.R
\name{geom_roc}
\alias{geom_roc}
\title{Geometry for ROC}
\usage{
geom_roc(data = NULL, mapping = aes(), size = 1, ...)
}
\arguments{
\item{data}{roc analysis result after 'pROC' or 'ROCit' package}

\item{mapping}{mapping}

\item{size}{line size defaule is 1}

\item{...}{other attributes}
}
\value{
gemeotry of roc
}
\description{
Geometry for ROC
}
\examples{
# do three ROC analysis
library(geomROC)
r1 <- rocit(score = Diabetes$bmi, class = Diabetes$dtest)
r2 <- rocit(score = Diabetes$chol, class = Diabetes$dtest)
r3 <- rocit(score = Diabetes$hdl, class = Diabetes$dtest)

#plot one roc curve
ggplot()+
    geom_roc(r1)
#plot two roc curves
ggplot()+
    geom_roc(r1)+
    geom_roc(r2)
#plot three roc curves
ggplot()+
    geom_roc(r1)+
    geom_roc(r2)+
    geom_roc(r3)

#use color
#add color to one roc
ggplot()+
    geom_roc(r1,aes(color='bmi'))+
    geom_roc(r2)+
    geom_roc(r3)
#add color to all
ggplot()+
    geom_roc(r1,aes(color='bmi'))+
    geom_roc(r2,aes(color='chol'))+
    geom_roc(r3,aes(color='hdl'))

#use linetype
ggplot()+
    geom_roc(r1,aes(linetype='bmi'))+
    geom_roc(r2,aes(linetype='chol'))+
    geom_roc(r3,aes(linetype='hdl'))

#use roc_theme()
ggplot()+
    geom_roc(r1,aes(linetype='bmi'))+
    geom_roc(r2,aes(linetype='chol'))+
    geom_roc(r3,aes(linetype='hdl'))+
    roc_theme()
}
