% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_aes.R
\name{roc-aes}
\alias{roc-aes}
\alias{roc_aes_color}
\alias{roc_aes_linetype}
\title{ROC Aesthetic}
\usage{
roc_aes_color(...)

roc_aes_linetype(...)
}
\arguments{
\item{...}{one or more colors or linetypes}
}
\description{
ROC Aesthetic
}
