% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal.R
\name{roc_diagonal}
\alias{roc_diagonal}
\title{Diagonal}
\usage{
roc_diagonal(mapping = aes(), color = "grey", size = 1.05,
  linetype = 2, ...)
}
\arguments{
\item{mapping}{mapping}

\item{color}{line color, black defaulted}

\item{size}{line size, 0.5 defaulted}

\item{linetype}{line type, numeric object, 2 is defaulted}

\item{...}{other function passed to params}
}
\value{
A diagonal
}
\description{
Diagonal
}
