% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_theme.R
\name{roc_theme}
\alias{roc_theme}
\title{Theme for ROC}
\usage{
roc_theme(axis_title_size = 14, axis_text_size = 12,
  legend_text_size = 12, legend_line_size = 20, axis_line_size = 1,
  axis_ticks_size = 0.5, border_size = 0, family = "")
}
\arguments{
\item{axis_title_size}{axis title size for x and y}

\item{axis_text_size}{axis text size for x and y}

\item{legend_text_size}{legend text size}

\item{legend_line_size}{legend key size, key type if line}

\item{axis_line_size}{axis line size for x and y}

\item{axis_ticks_size}{axis ticks size for x and y}

\item{border_size}{border size}

\item{family}{family}
}
\description{
Theme for ROC
}
